dont pass down viewmodelScope
